requirements.txt
requests
psutil
pre-commit
PySide6
black
Matplotlib
pyinstaller
python-dotenv
pytest
pytest-mock
aiosqlite
aiofiles
asyncio
aiohttp
pytest-asyncio
qasync
ipinfo
